Apply redux async action concept and fetch api mechanism to entity function
