+++ entity_panel/containers/EntityControl.js
@@ -1,13 +1,13 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { addEntity, saveEntities } from '../../common/actions'
+import { addEntity, submitEntities } from '../../common/actions'
 let EntityControl = ({ dispatch }) => (
   <div>
     <button className="btn btn-default"
       onClick={() => dispatch(addEntity())}>Add Entity
     </button>
     <button className="btn btn-default"
-      onClick={() => dispatch(saveEntities())}>Save
+      onClick={() => dispatch(submitEntities())}>Submit
     </button>
   </div>
 )
